Split & reorder messengers
