+++ src/controller/main.go
@@ -101,7 +101,6 @@ func initMessengers() (map[string]Messenger, error) {
 }
 
 func newHTTPListener() net.Listener {
-
 	lis, err := net.Listen("tcp", fmt.Sprintf(":%s", config.ServerPort))
 	if err != nil {
 		log.Fatalf("failed to listen: %v", err)
