+++ src/messengers/vk/service.go
@@ -17,10 +17,10 @@ func main() {
 	if err != nil {
 		log.Fatalf("failed to listen: %v", err)
 	}
-	messenger := createMessenger()
-	go messenger.Run(context.Background())
+	vkMessenger := createMessenger()
+	go vkMessenger.Run(context.Background())
 	grpcServer := grpc.NewServer()
-	msg.RegisterChatServiceServer(grpcServer, messenger)
+	msg.RegisterChatServiceServer(grpcServer, vkMessenger)
 
 	log.Printf("initializing gRPC server on port %d", vk.Config.Port)
 	if err = grpcServer.Serve(lis); err != nil {
@@ -28,12 +28,15 @@ func main() {
 	}
 }
 
-func createMessenger() *vk.VKMessenger {
+func createMessenger() *vk.Messenger {
 	connection, err := grpc.Dial(vk.Config.ControllerHost, grpc.WithTransportCredentials(insecure.NewCredentials()))
 	if err != nil {
 		log.Fatalf("could not connect to controller on %s", vk.Config.ControllerHost)
 	}
-	vkMessenger := vk.NewVKMessenger(messenger.NewBaseMessenger(connection))
+	vkMessenger, err := vk.NewMessenger(messenger.NewBaseMessenger(connection))
+	if err != nil {
+		log.Fatalf("could not create messenger: %v", err)
+	}
 	log.Printf("connected to controller on %s", vk.Config.ControllerHost)
 	return vkMessenger
 }
